feat: create API for displaying a list of publishers in pagination form

+++ src/main/java/com/gustionusamba/bookcatalog/dto/ResultPageResponseDTO.java
@@ -17,7 +17,7 @@ public class ResultPageResponseDTO<T> implements Serializable {
 
     private List<T> result;
 
-    private Long pages;
+    private Integer pages;
 
     private Long elements;
 }
